+++ src/score/engine/obj-bar-line.ts
@@ -2,7 +2,7 @@ import { MusicObject } from "./music-object";
 import { View } from "./view";
 import { ObjMeasure } from "./obj-measure";
 import { MBarLineRight, MBarLineLeft, Navigation, MusicInterface, MStaffBarLine } from "../pub";
-import { PlayerColumnProps } from "./player";
+import { PlayerColumnProps } from "./player-engine";
 import { DocumentSettings } from "./settings";
 import { ObjNotationLine, ObjStaff } from "./obj-staff-and-tab";
 import { AnchoredRect, UniMap } from "@tspro/ts-utils-lib";